moved Profile in the ProfileComponents folder
