+++ src/Web/src/components/ProfileComponents/Profile.js
@@ -1,17 +1,18 @@
 import React, { useState } from 'react';
 import { Nav, NavItem, NavLink} from 'reactstrap';
 
-import ProfileInfo from './ProfileComponents/ProfileInfo';
-import CollapsibleTable from './ProfileComponents/CollapsibleTable';
-import LeaderOfOrgChart from './ProfileComponents/LeaderOfOrgChart';
+import ProfileInfo from './ProfileInfo';
+import CollapsibleTable from './CollapsibleTable';
+import LeaderOfOrgChart from './LeaderOfOrgChart';
+import UseProfile from './UseProfile';
 
 const Profile = () => {
     const [activeComponent, setActiveComponent] = useState("general");
+    const { data } = UseProfile();
 
     return (
         <div>
-            <ProfileInfo />
-
+            <ProfileInfo data={data} />
             <Nav className="profile-nav">
                 <NavItem className={activeComponent === "general" ? "current-profile-page nav-option" : "nav-option"}>
                     <NavLink onClick={() => setActiveComponent("general")}>General Info</NavLink>
@@ -21,12 +22,12 @@ const Profile = () => {
                 </NavItem>
             </Nav>
 
-            { activeComponent === "general" ? (
+            { activeComponent === "general" && data !== {} ? (
                 <div>
-                    <CollapsibleTable title='Education' categories={['Institution', 'Degree', 'Date', 'Specialization']} />
-                    <CollapsibleTable title='Position History' categories={['Role', 'School', 'Start Date', 'End Date']} />
-                    <CollapsibleTable title='Certifications' categories={['Description', 'Tyoe', 'Valid from', 'Valid to']} />
-                    <CollapsibleTable title='Professional Development and Learning Experiences' categories={['Course name', 'Date', 'Location', 'Alignment to leadership definition']} />
+                    <CollapsibleTable title='Education' data={data.education} />
+                    <CollapsibleTable title='Position History' data={data.positionHistory} />
+                    <CollapsibleTable title='Certifications' data={data.certificates} />
+                    <CollapsibleTable title='Professional Development and Learning Experiences' data={data.professionalDevelopment}/>
                 </div>
                 ) : activeComponent === "leader" ? (
                     <LeaderOfOrgChart />
